p-n-p
